pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-sync"
version = "0.1.0"
description = "One-way sync from Things 3 and Asana into Taskwarrior"
requires-python = ">=3.11"
dependencies = [
    "things.py>=0.0.15",
    "asana>=5.0.0",
    "tasklib>=2.5.1",
    "click>=8.1",
    "tomli>=2.0;python_version<'3.11'",
]

[project.scripts]
todo-sync = "todo_sync.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
